修复Zip64处理的bug 修复循环定时任务处理的bug 加入QOI5551格式编码器，适用于嵌入式16位液晶屏 加入SegmentedSource，以offset和length分割一个Source
Lavac
修复自增表达式的转换与求值问题 Assign#incOrDec
修复错误的英文翻译 inv nullish 等
现在可以手动生成枚举特定方法 CompileUnit#L1726
加入errorReportingIndex提高代码可读性
修复LocalContext#inferGeneric的推导bug，同时改进它的javadoc
修复Asterisk类型在推导时的嵌套问题
加入通用静态变量DFI
使用方法体注解替代之前的特殊token来禁用优化

+++ java/roj/archive/zip/ZEntry.java
@@ -222,15 +222,15 @@ public class ZEntry implements RSegmentTree.Range, ArchiveEntry {
 					}
 					break;
 				case 0x0001: // Zip64 extended information
-					if (uSize == (int) U32_MAX && len >= 8) {
+					if (uSize == U32_MAX && len >= 8) {
 						len -= 8;
 						uSize = buf.readLongLE();
 					}
-					if (cSize == (int) U32_MAX && len >= 8) {
+					if (cSize == U32_MAX && len >= 8) {
 						len -= 8;
 						cSize = buf.readLongLE();
 					}
-					if (header == (int) U32_MAX && len >= 8) {
+					if (header == U32_MAX && len >= 8) {
 						header = buf.readLongLE();
 					}
 					break;
